+++ requirements.txt
@@ -1,0 +1,2 @@
+bs4
+colorama
